refactor: extract Presets and Settings from TemperaturePanel
